[Fix] some css and errors in the UI and move the random-meal code to it's only file
