+++ src/js/utils.js
@@ -1,11 +1,6 @@
-const mealModalTemplate = document
-  .querySelector('.meal')
-  .content.querySelector('.modal')
-  .cloneNode(true);
-const body = document.body;
-
 import arrowDownIcon from '../assets/icons/load-more.svg';
 import { renderCategories, renderMealsByCategory } from './category.js';
+import { getMeal } from './randomMeal.js';
 
 // API information
 export const API = {
@@ -56,7 +51,8 @@ export function createMealCard(meal) {
 
 // Render more meals
 export function renderMoreMeals(restList, container, event) {
-  event.target.parentElement.remove();
+  const parentLoadMore = event.target.parentElement.parentElement;
+  parentLoadMore.remove();
   const fragment = document.createDocumentFragment();
   restList
     .slice(0, NUM_MEALS_TO_LOAD)
@@ -83,83 +79,4 @@ export function addLodeMoreBtn(refContainer, restMeals) {
   refContainer.appendChild(addMoreTemplate);
 }
 
-export const getRandomMeal = async () => {
-  const randomUrl = `${API.base}${API.random}`;
-  const randomMeal = (await getDataInJson(randomUrl)).meals[0];
 
-  generateModal(randomMeal);
-
-  return randomMeal;
-};
-
-export const getMeal = async (meal) => {
-  const url = `${API.base}${API.lookUp}?i=${meal.idMeal}`;
-  const retrievedMeal = (await getDataInJson(url)).meals[0];
-
-  generateModal(retrievedMeal);
-
-  return retrievedMeal;
-};
-
-export const generateModal = (meal) => {
-  mealModalTemplate.querySelector(
-    '.meal__image'
-  ).style.backgroundImage = `url('${meal.strMealThumb}`;
-  mealModalTemplate.querySelector('.meal__name').textContent = meal.strMeal;
-  mealModalTemplate.querySelector(
-    '.meal__preparation-instructions'
-  ).textContent = meal.strInstructions;
-
-  const ingredientContainer =
-    mealModalTemplate.querySelector('.ingredients__list');
-  const ingredients = getIngredients(meal).map((i) => {
-    return i.replace(/^\s+|\s+$/g, '') !== '' && i !== ' null' && i !== 'null null'
-      ? i
-      : undefined;
-  });
-  
-  ingredientContainer.innerHTML = '';
-  ingredients.forEach((ingredient) => {
-    if (ingredient !== undefined) {
-      let ingredientLi = document.createElement('li');
-      ingredientLi.textContent = ingredient;
-      ingredientContainer.appendChild(ingredientLi);
-    }
-  });
-
-  const closeButton = mealModalTemplate.querySelector('.close__button');
-
-  closeButton.addEventListener('click', (e) => {
-    e.preventDefault();
-
-    e.target.parentElement.parentElement.parentElement.parentElement.remove();
-  });
-
-  body.append(mealModalTemplate);
-};
-
-export const getIngredients = (meal) => {
-  const ingredients = [
-    `${meal.strMeasure1} ${meal.strIngredient1}`,
-    `${meal.strMeasure2} ${meal.strIngredient2}`,
-    `${meal.strMeasure3} ${meal.strIngredient3}`,
-    `${meal.strMeasure4} ${meal.strIngredient4}`,
-    `${meal.strMeasure5} ${meal.strIngredient5}`,
-    `${meal.strMeasure6} ${meal.strIngredient6}`,
-    `${meal.strMeasure7} ${meal.strIngredient7}`,
-    `${meal.strMeasure8} ${meal.strIngredient8}`,
-    `${meal.strMeasure9} ${meal.strIngredient9}`,
-    `${meal.strMeasure10} ${meal.strIngredient10}`,
-    `${meal.strMeasure11} ${meal.strIngredient11}`,
-    `${meal.strMeasure12} ${meal.strIngredient12}`,
-    `${meal.strMeasure13} ${meal.strIngredient13}`,
-    `${meal.strMeasure14} ${meal.strIngredient14}`,
-    `${meal.strMeasure15} ${meal.strIngredient15}`,
-    `${meal.strMeasure16} ${meal.strIngredient16}`,
-    `${meal.strMeasure17} ${meal.strIngredient17}`,
-    `${meal.strMeasure18} ${meal.strIngredient18}`,
-    `${meal.strMeasure19} ${meal.strIngredient19}`,
-    `${meal.strMeasure20} ${meal.strIngredient20}`,
-  ];
-  return ingredients;
-};
